requirements.txt
opencv-python
stegano
pycryptodome
Pillow
termcolor
pyfiglet
numpy
